// web/src/api/generated/model/stockDocumentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { StockDocumentKind } from './stockDocumentKind';
import type { StockLineInput } from './stockLineInput';

export interface StockDocumentInput {
  kind: StockDocumentKind;
  date: string;
  /** @nullable */
  memo: string | null;
  /** @nullable */
  counterAccountId: string | null;
  lines: StockLineInput[];
}