// web/src/api/generated/model/stockDocumentKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type StockDocumentKind = typeof StockDocumentKind[keyof typeof StockDocumentKind];


export const StockDocumentKind = {
  Opening: 'Opening',
  Adjustment: 'Adjustment',
  Transfer: 'Transfer',
} as const;