// web/src/api/generated/model/stockLineInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface StockLineInput {
  productId: string;
  warehouseId: string;
  /** @nullable */
  toWarehouseId: string | null;
  quantity: number;
  /** @nullable */
  unitCost: number | null;
}